Master --- name: Pull request about: Fix issues in a lesson's code title: "" labels: bug assignees: "" ---
Thanks for making this PR! Future learners thank you for your diligence. In
order to help the curriculum team address the problem, please use this template
to submit your feedback. We'll work on addressing the issue as soon as we can.

Please fill out as much of the information below as you can (it's ok if you
don't fill out every section). The more context we have, the easier it will be
to fix your issue!

---

## Link to Canvas

Add a link to the assignment in Canvas here.

## What changes does this PR make?

- Item 1
- Item 2

## What issue does this PR fix?

Describe the issue here, or [reference an open issue](https://github.blog/2013-05-14-closing-issues-via-pull-requests/) that this PR fixes.

- Fixes #999